--- frontend/src/App.jsx
@@ -1,6 +1,0 @@
-import Layout from './components/Layout'
-import './App.css'
-
-export default function App(){
-    return <Layout/>
-}
