+++ frontend/src/App/Map/index.jsx
@@ -1,9 +1,9 @@
 import { Component } from 'react'
 import mapboxgl from 'mapbox-gl'
 import './Mapbox.css'
-import { getClosestNode, updateClosestNode } from '../../actions'
-import arrowImage from '../Images/arrow.png'
-import doubleArrowImage from '../Images/doublearrow.png'
+import { getClosestNode, updateClosestNode } from '../../actions.js'
+import arrowImage from './images/arrow.png'
+import doubleArrowImage from './images/doublearrow.png'
 import { NotificationContainer, NotificationManager } from 'react-notifications'
 import ActionsBox from './ActionsBox'
 import 'react-notifications/lib/notifications.css'
