+++ frontend/src/App/Sidebar/SidebarContent.css
