+++ frontend/src/App/app.css
@@ -58,3 +58,17 @@
     bottom: 0;
   
 }
+
+.header{
+    background-color: black;
+}
+
+.sidebar{
+    width: 490px;
+    background-color: #ffffff;
+}
+
+.topbar{
+    height: 100%;
+    background-color: #1a1a1a;
+}
