+++ frontend/src/actions.js
@@ -4,7 +4,7 @@ const axios = require('axios');
 axios.defaults.withCredentials = true;
 /* remote domain and local test domain */
 const domain = process.env.NODE_ENV === 'development' ? 
-    'http://127.0.0.1:5000' : 'https://10.160.2.198/tt-request-backend';
+    'http://127.0.0.1:5001' : 'https://10.160.2.198/tt-request-backend';
 const fileDownload = require('js-file-download');
 
 const handleResponseError = (err) => {
