--- frontend/src/components/Layout/Layout.css
@@ -1,13 +1,0 @@
-.header{
-    background-color: black;
-}
-
-.sidebar{
-    width: 490px;
-    background-color: #ffffff;
-}
-
-.topbar{
-    height: 100%;
-    background-color: #1a1a1a;
-}
